# generated/backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DDF 精密五金 MES API",
    "version": "1.0.0",
    "description": "由 DomainDriveFramework 引擎自动生成",
    "contact": {
      "name": "DDF Engine"
    }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "本地开发" }
  ],
  "paths": {
...
    "/api/material-category": {
      "get": {
        "summary": "分页查询物料分类",
        "tags": ["MaterialCategory"],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string" } },
          { "name": "sortOrder", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "分页结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/MaterialCategory" } },
                    "total": { "type": "integer" },
                    "page": { "type": "integer" },
                    "pageSize": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建物料分类",
        "tags": ["MaterialCategory"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MaterialCategoryCreate" }
            }
          }
        },
        "responses": {
          "201": { "description": "创建成功" }
        }
      }
    },
    "/api/material-category/{id}": {
      "get": {
        "summary": "查询物料分类详情",
        "tags": ["MaterialCategory"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "实体详情" }
        }
      },
      "put": {
        "summary": "更新物料分类",
        "tags": ["MaterialCategory"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MaterialCategoryCreate" }
            }
          }
        },
        "responses": {
          "200": { "description": "更新成功" }
        }
      },
      "delete": {
        "summary": "删除物料分类",
        "tags": ["MaterialCategory"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "删除成功" }
        }
      }
    },
    "/api/material": {
      "get": {
        "summary": "分页查询物料",
        "tags": ["Material"],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string" } },
          { "name": "sortOrder", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "分页结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Material" } },
                    "total": { "type": "integer" },
                    "page": { "type": "integer" },
                    "pageSize": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建物料",
        "tags": ["Material"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MaterialCreate" }
            }
          }
        },
        "responses": {
          "201": { "description": "创建成功" }
        }
      }
    },
    "/api/material/{id}": {
      "get": {
        "summary": "查询物料详情",
        "tags": ["Material"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "实体详情" }
        }
      },
      "put": {
        "summary": "更新物料",
        "tags": ["Material"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MaterialCreate" }
            }
          }
        },
        "responses": {
          "200": { "description": "更新成功" }
        }
      },
      "delete": {
        "summary": "删除物料",
        "tags": ["Material"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "删除成功" }
        }
      }
    },
...
    "/api/material-attr-value": {
      "get": {
        "summary": "分页查询物料属性值",
        "tags": ["MaterialAttrValue"],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string" } },
          { "name": "sortOrder", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "分页结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/MaterialAttrValue" } },
                    "total": { "type": "integer" },
                    "page": { "type": "integer" },
                    "pageSize": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建物料属性值",
        "tags": ["MaterialAttrValue"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MaterialAttrValueCreate" }
            }
          }
        },
        "responses": {
          "201": { "description": "创建成功" }
        }
      }
    },
    "/api/material-attr-value/{id}": {
      "get": {
        "summary": "查询物料属性值详情",
        "tags": ["MaterialAttrValue"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "实体详情" }
        }
      },
      "put": {
        "summary": "更新物料属性值",
        "tags": ["MaterialAttrValue"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MaterialAttrValueCreate" }
            }
          }
        },
        "responses": {
          "200": { "description": "更新成功" }
        }
      },
      "delete": {
        "summary": "删除物料属性值",
        "tags": ["MaterialAttrValue"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "删除成功" }
        }
      }
    },
...
    "/api/work-order": {
      "get": {
        "summary": "分页查询工单",
        "tags": ["WorkOrder"],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string" } },
          { "name": "sortOrder", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "分页结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/WorkOrder" } },
                    "total": { "type": "integer" },
                    "page": { "type": "integer" },
                    "pageSize": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建工单",
        "tags": ["WorkOrder"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WorkOrderCreate" }
            }
          }
        },
        "responses": {
          "201": { "description": "创建成功" }
        }
      }
    },
    "/api/work-order/{id}": {
      "get": {
        "summary": "查询工单详情",
        "tags": ["WorkOrder"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "实体详情" }
        }
      },
      "put": {
        "summary": "更新工单",
        "tags": ["WorkOrder"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WorkOrderCreate" }
            }
          }
        },
        "responses": {
          "200": { "description": "更新成功" }
        }
      },
      "delete": {
        "summary": "删除工单",
        "tags": ["WorkOrder"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "删除成功" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MaterialCategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "category_code": {
            "type": "string"
          },
          "category_name": {
            "type": "string"
          },
          "level": {
            "type": "integer"
          },
          "parent_id": {
            "type": "integer"
          }
        }
      },
      "MaterialCategoryCreate": {
        "type": "object",
        "properties": {
          "category_code": {
            "type": "string",
            "nullable": false
          },
          "category_name": {
            "type": "string",
            "nullable": false
          },
          "level": {
            "type": "integer",
            "nullable": false
          },
          "parent_id": {
            "type": "integer",
            "nullable": false
          }
        }
      },
      "Material": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "material_code": {
            "type": "string"
          },
          "material_name": {
            "type": "string"
          },
          "specification": {
            "type": "string"
          },
          "category_id": {
            "type": "integer"
          },
          "unit": {
            "type": "string"
          },
          "material_type": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
      "MaterialCreate": {
        "type": "object",
        "properties": {
          "material_code": {
            "type": "string",
            "nullable": false
          },
          "material_name": {
            "type": "string",
            "nullable": false
          },
          "specification": {
            "type": "string",
            "nullable": false
          },
          "category_id": {
            "type": "integer",
            "nullable": false
          },
          "unit": {
            "type": "string",
            "nullable": false
          },
          "material_type": {
            "type": "string",
            "nullable": false
          },
          "status": {
            "type": "string",
            "nullable": false
          }
        }
      },
...
      "MaterialAttrValue": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "material_id": {
            "type": "integer"
          },
          "attr_id": {
            "type": "integer"
          },
          "value": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          }
        }
      },
      "MaterialAttrValueCreate": {
        "type": "object",
        "properties": {
          "material_id": {
            "type": "integer",
            "nullable": false
          },
          "attr_id": {
            "type": "integer",
            "nullable": false
          },
          "value": {
            "type": "string",
            "nullable": false
          }
        }
      },
...
      "WorkOrder": {
        "type": "object",
        "properties": {
          "order_no": {
            "type": "string"
          },
          "product_code": {
            "type": "string"
          },
          "product_name": {
            "type": "string"
          },
          "quantity": {
            "type": "integer"
          },
          "completed_quantity": {
            "type": "integer"
          },
          "scrap_quantity": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "plan_start_date": {
            "type": "string"
          },
          "plan_end_date": {
            "type": "string"
          },
          "actual_start_time": {
            "type": "string"
          },
          "actual_end_time": {
            "type": "string"
          },
          "route_id": {
            "type": "integer"
          },
          "production_line_id": {
            "type": "integer"
          },
          "shift_id": {
            "type": "integer"
          },
          "comment": {
            "type": "string"
          },
          "created_by": {
            "type": "integer"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
      "WorkOrderCreate": {
        "type": "object",
        "properties": {
          "order_no": {
            "type": "string",
            "nullable": false
          },
          "product_code": {
            "type": "string",
            "nullable": false
          },
          "product_name": {
            "type": "string",
            "nullable": false
          },
          "quantity": {
            "type": "integer",
            "nullable": false
          },
          "completed_quantity": {
            "type": "integer",
            "nullable": false
          },
          "scrap_quantity": {
            "type": "integer",
            "nullable": false
          },
          "status": {
            "type": "string",
            "nullable": false
          },
          "priority": {
            "type": "string",
            "nullable": false
          },
          "plan_start_date": {
            "type": "string",
            "nullable": false
          },
          "plan_end_date": {
            "type": "string",
            "nullable": false
          },
          "actual_start_time": {
            "type": "string",
            "nullable": false
          },
          "actual_end_time": {
            "type": "string",
            "nullable": false
          },
          "route_id": {
            "type": "integer",
            "nullable": false
          },
          "production_line_id": {
            "type": "integer",
            "nullable": false
          },
          "shift_id": {
            "type": "integer",
            "nullable": false
          },
          "comment": {
            "type": "string",
            "nullable": false
          }
        }
      },
...
    }
  }
}