# generated/backend/routes/work-order.routes.js
/**
 * WorkOrder CRUD 路由
 * Domain: base.work_order
 * Generated by DDF Engine
 */

const express = require('express');
const router = express.Router();
const prisma = require('../prisma/client');

// 字段白名单（排除审计字段）
const FIELDS = ['order_no', 'product_code', 'product_name', 'quantity', 'completed_quantity', 'scrap_quantity', 'status', 'priority', 'plan_start_date', 'plan_end_date', 'actual_start_time', 'actual_end_time', 'route_id', 'production_line_id', 'shift_id', 'comment'];

// 可排序字段列表（由 DDF Engine 在生成期计算）
const SORTABLE_FIELDS = ['order_no', 'product_code', 'product_name', 'quantity', 'completed_quantity', 'scrap_quantity', 'plan_start_date', 'plan_end_date', 'actual_start_time', 'actual_end_time', 'comment', 'created_at', 'updated_at'];

// 可搜索字段列表（由 DDF Engine 在生成期计算）
const SEARCHABLE_FIELDS = ['order_no', 'product_code', 'product_name', 'status', 'priority', 'comment'];

// ---- 字段值映射函数 ----
function mapInput(body) {
  const data = {};
  for (const key of FIELDS) {
    if (body[key] !== undefined) {
      data[key] = body[key];
    }
  }
  return data;
}

/**
 * GET /api/work-order
 * 列表查询（分页 + 排序 + 搜索）
 */
router.get('/', async (req, res) => {
  try {
    const { page = 1, pageSize = 20, sortBy, sortOrder, search, ...filters } = req.query;
    const skip = (parseInt(page) - 1) * parseInt(pageSize);
    const take = parseInt(pageSize);

    const where = { ...filters };

    // 全文搜索
    if (search) {
      where.OR = SEARCHABLE_FIELDS.map(field => ({
        [field]: { contains: search },
      }));
    }

    // 排序（仅允许在 SORTABLE_FIELDS 范围内）
    let orderBy = { id: 'desc' };
    if (sortBy && SORTABLE_FIELDS.includes(sortBy)) {
      orderBy = { [sortBy]: sortOrder || 'asc' };
    }

    const [data, total] = await Promise.all([
      prisma.workOrder.findMany({ where, skip, take, orderBy }),
      prisma.workOrder.count({ where }),
    ]);

    res.json({ data, total, page: parseInt(page), pageSize: parseInt(pageSize) });
  } catch (err) {
    res.status(500).json({ error: err.message });
  }
});

/**
 * GET /api/work-order/:id
 * 详情
 */
router.get('/:id', async (req, res) => {
  try {
    const data = await prisma.workOrder.findUnique({
      where: { id: parseInt(req.params.id) },
    });
    if (!data) return res.status(404).json({ error: 'Not found' });
    res.json(data);
  } catch (err) {
    res.status(500).json({ error: err.message });
  }
});

/**
 * POST /api/work-order
 * 创建
 */
router.post('/', async (req, res) => {
  try {
    const data = await prisma.workOrder.create({
      data: mapInput(req.body),
    });
    res.status(201).json(data);
  } catch (err) {
    res.status(400).json({ error: err.message });
  }
});

/**
 * PUT /api/work-order/:id
 * 更新
 */
router.put('/:id', async (req, res) => {
  try {
    const data = await prisma.workOrder.update({
      where: { id: parseInt(req.params.id) },
      data: mapInput(req.body),
    });
    res.json(data);
  } catch (err) {
    res.status(400).json({ error: err.message });
  }
});

/**
 * DELETE /api/work-order/:id
 * 删除
 */
router.delete('/:id', async (req, res) => {
  try {
    await prisma.workOrder.delete({
      where: { id: parseInt(req.params.id) },
    });
    res.status(204).send();
  } catch (err) {
    res.status(400).json({ error: err.message });
  }
});

module.exports = router;